+++ web/config-overrides.js
@@ -7,6 +7,6 @@ module.exports = override(
   }),
   addLessLoader({
     javascriptEnabled: true,
-    modifyVars: { "@primary-color": "#1DA57A" },
+    // modifyVars: { "@primary-color": "#1DA57A" },
   })
 );
