+++ web/src/components/Main/index.css
@@ -1,0 +1,3 @@
+.site-layout-background {
+  background: #fff;
+}
