+++ web/src/pages/User/index.jsx
@@ -1,0 +1,5 @@
+const User = (props) => {
+  return <div>user</div>;
+};
+
+export default User;
