--- web/src/pages/components/Aside/index.jsx
@@ -1,45 +1,0 @@
-import { NavLink, withRouter } from "react-router-dom";
-import { Layout, Menu } from "antd";
-import {
-  CalendarOutlined,
-  DashboardOutlined,
-  WechatOutlined,
-  FireOutlined,
-} from "@ant-design/icons";
-
-const { Sider } = Layout;
-function Aside(props) {
-  const { pathname } = props.location;
-  const keyArray = pathname.split("/");
-  const selectedKey = keyArray[1] === undefined ? "dashboard" : keyArray[1];
-  return (
-    <Sider width={200} className="site-layout-background">
-      <Menu
-        // theme="dark"
-        mode="inline"
-        defaultSelectedKeys={["dashboard"]}
-        selectedKeys={[selectedKey]}
-        style={{ height: "100%", borderRight: 0 }}
-      >
-        <Menu.Item key="dashboard" icon={<DashboardOutlined />}>
-          Dash Board
-          <NavLink to="/dashboard" />
-        </Menu.Item>
-        <Menu.Item key="discussion" icon={<FireOutlined />}>
-          Discussion Board
-          <NavLink to="/discussion" />
-        </Menu.Item>
-        <Menu.Item key="chat" icon={<WechatOutlined />}>
-          Chat Room
-          <NavLink to="/chat" />
-        </Menu.Item>
-        <Menu.Item key="Calendar" icon={<CalendarOutlined />}>
-          Calendar Schedule
-          <NavLink to="/Calendar" />
-        </Menu.Item>
-      </Menu>
-    </Sider>
-  );
-}
-
-export default withRouter(Aside);
