--- web/src/pages/components/GroupLive/index.css
@@ -1,9 +1,0 @@
-.logo {
-  float: left;
-  height: 31px;
-  margin: 16px 0;
-}
-
-.site-layout-background {
-  background: #fff;
-}
